requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
easyocr==1.7.2
fastapi==0.136.1
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.3.0
h11==0.16.0
idna==3.13
ImageIO==2.37.3
Jinja2==3.1.6
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.1
mysql-connector-python==9.6.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
opencv-python-headless==4.13.0.92
packaging==26.2
pillow==12.2.0
propcache==0.4.1
pyclipper==1.4.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
python-bidi==0.6.7
python-dotenv==1.2.2
python-multipart==0.0.27
PyYAML==6.0.3
pyzbar==0.1.9
qrcode==8.2
requests==2.33.1
scikit-image==0.26.0
scipy==1.17.1
setuptools==81.0.0
shapely==2.1.2
starlette==1.0.0
sympy==1.13.1
tifffile==2026.4.11
torch==2.6.0+cu124
torchvision==0.21.0+cu124
twilio==9.10.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uvicorn==0.46.0
yarl==1.23.0
